requirements.txt
Cython==0.23.4
Flask==0.10.1
Jinja2==2.7.2
Mako==1.0.3
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.11
Twisted-Core==13.2.0
Twisted-Web==13.2.0
UNKNOWN==0.0.0
Werkzeug==0.11.2
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
apt-xapian-index==0.45
argparse==1.2.1
autokey==0.90.4
awscli==1.9.9
awsebcli==3.6.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.2.1
blessed==1.9.5
boto==2.38.0
botocore==1.3.9
bottle==0.12.9
bz2file==0.98
cement==2.4.0
chainer==1.5.0.2
chainer-cuda-deps==1.1.0.1
chardet==2.0.1
colorama==0.3.3
command-not-found==0.3
cvxopt==1.1.4
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
duplicity==0.6.23
filelock==2.0.5
gensim==0.12.3
gyp==0.1
h5py==2.5.0
html5lib==0.999
httplib2==0.8
httpretty==0.8.6
ipython==2.3.0
itsdangerous==0.24
jmespath==0.9.0
joblib==0.7.1
kenlm==0.0.0
lockfile==0.8
lxml==3.3.3
matplotlib==1.3.1
nltk==3.1
nose==1.3.1
numexpr==2.2.2
numpy==1.10.1
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
openpyxl==1.7.0
pandas==0.13.1
pathspec==0.3.3
patsy==0.2.1
pexpect==3.1
piston-mini-client==0.7.5
protobuf==2.6.1
py==1.4.31
pyOpenSSL==0.13
pyasn1==0.1.9
pycrypto==2.6.1
pycuda==2015.1.3
pycups==1.9.66
pycurl==7.19.3
pyenchant==1.6.6
pygobject==3.12.0
pyinotify==0.9.4
pyparsing==2.0.1
pyserial==2.6
pysmbc==1.0.14.1
pytest==2.8.3
python-apt==0.9.3.5ubuntu1
python-dateutil==2.4.2
python-debian==0.1.21-nmu2ubuntu2
python-xlib==0.14
pytools==2015.1.6
pytz==2012c
pyxdg==0.25
pyzmq==14.0.1
reportlab==3.0
requests==2.6.2
rsa==3.2.3
scikit-cuda==0.5.1
scikit-learn==0.16.1
scipy==0.13.3
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
smart-open==1.3.0
software-center-aptd-plugins==0.0.0
statsmodels==0.5.0
sympy==0.7.4.1
system-service==0.1.6
tables==3.1.1
tensorflow==0.5.0
texttable==0.8.4
tornado==3.1.1
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==13.1.2
wcwidth==0.1.5
web.py==0.37
websocket-client==0.34.0
wget==3.2
wheel==0.24.0
wikipedia==1.4.0
word2vec==v0.8.1
wsgiref==0.1.2
xdiagnose==3.6.3build2
xlrd==0.9.2
xlwt==0.7.5
zope.interface==4.0.5
